Let users import a saved project from the CNT page too
